Move namespace around
