+++ soulmemory-rs/src/games/dark_souls_remastered.rs
@@ -22,9 +22,10 @@ use retour::static_detour;
 use log::info;
 use mem_rs::prelude::*;
 use windows::Win32::UI::Input::XboxController::XINPUT_STATE;
-use crate::{App, GameExt};
+use crate::{App};
 use crate::games::dx_version::DxVersion;
-use crate::games::traits::game::Game;
+use crate::games::game::Game;
+use crate::games::GameExt;
 use crate::tas::tas::{get_xinput_get_state_fn, tas_ai_toggle};
 use crate::tas::toggle_mode::ToggleMode;
 use crate::games::traits::buffered_event_flags::{BufferedEventFlags, EventFlag};
