+++ soulmemory-rs/src/games/mod.rs
@@ -15,14 +15,30 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 pub mod traits;
-pub mod dark_souls_remastered;
-pub mod dark_souls_prepare_to_die_edition;
-pub mod dark_souls_3;
-pub mod sekiro;
-pub mod elden_ring;
-pub mod armored_core_6;
-pub mod mock_game;
+mod dark_souls_remastered;
+mod dark_souls_prepare_to_die_edition;
+mod dark_souls_3;
+mod sekiro;
+mod elden_ring;
+mod armored_core_6;
+mod mock_game;
 pub mod dx_version;
+mod game;
+mod game_ext;
+
+pub use game::Game;
+pub use game_ext::GameExt;
+
+
+pub use mock_game::MockGame;
+pub use dark_souls_remastered::DarkSoulsRemastered;
+pub use dark_souls_prepare_to_die_edition::DarkSoulsPrepareToDieEdition;
+pub use dark_souls_3::DarkSouls3;
+pub use sekiro::Sekiro;
+pub use elden_ring::EldenRing;
+pub use armored_core_6::ArmoredCore6;
+
+
 
 pub type ChrDbgFlag = (u32, String, bool);
