+++ soulmemory-rs/src/games/traits/mod.rs
@@ -1,4 +1,2 @@
-pub mod game;
-pub mod game_ext;
 pub mod player_position;
 pub mod buffered_event_flags;
