+++ soulmemory-rs/src/widgets/chr_dbg_flags_widget.rs
@@ -17,9 +17,7 @@
 use std::ops::DerefMut;
 use imgui::{TreeNodeFlags, Ui};
 use crate::games::{GetSetChrDbgFlags};
-use crate::games::sekiro::Sekiro;
-use crate::games::traits::game::{Game};
-use crate::games::traits::game_ext::GameExt;
+use crate::games::*;
 use crate::widgets::widget::Widget;
 
 pub struct ChrDbgFlagsWidget
